refactor: Move iObj functions in Toolkit back to iObj Class
