+++ lib/Toolkit.js
@@ -519,862 +519,14 @@ class Toolkit {
     this.conn.run(toJson); // Post the input XML and get the response.
   }
 
-  retrUsrAuth(usr, type, obj, lib = '*LIBL', cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // Object authority / Data authority
-        { type: '1A', value: '' }, // Authorization list management
-        { type: '1A', value: '' }, // Object operational
-        { type: '1A', value: '' }, // Object management
-        { type: '1A', value: '' }, // Object existence
-        { type: '1A', value: '' }, // Data read
-        { type: '1A', value: '' }, // Data add
-        { type: '1A', value: '' }, // Data update
-        { type: '1A', value: '' }, // Data delete
-        { type: '10A', value: '' }, // Authorization list
-        { type: '2A', value: '' }, // Authority source
-        { type: '1A', value: '' }, // Some adopted authority
-        { type: '10A', value: '' }, // Adopted object authority
-        { type: '1A', value: '' }, // Adopted authorization list management
-        { type: '1A', value: '' }, // Adopted object operational
-        { type: '1A', value: '' }, // Adopted object management
-        { type: '1A', value: '' }, // Adopted object existence
-        { type: '1A', value: '' }, // Adopted data read
-        { type: '1A', value: '' }, // Adopted data add
-        { type: '1A', value: '' }, // Adopted data update
-        { type: '1A', value: '' }, // Adopted data delete
-        { type: '1A', value: '' }, // Adopted data execute
-        { type: '10A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // Adopted object alter
-        { type: '1A', value: '' }, // Adopted object reference
-        { type: '10A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // Data execute
-        { type: '10A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // Object alter
-        { type: '1A', value: '' }, // Object reference
-        { type: '10A', value: '' }, // ASP device name of library
-        { type: '10A', value: '' }, // ASP device name of object
-        { type: '3A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Offset to group information table
-        { type: '10i0', value: 0 }, // Number of group table entries returned
-        { type: '48b', value: 0 }, // Group information table repeated for each of the user's groups
-      ],
-    };
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: obj },
-        { type: '10A', value: lib },
-      ],
-    };
-    const pgm = new ProgramCall('QSYRUSRA', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'USRA0100' });
-    pgm.addParam({ type: '10A', value: usr });
-    pgm.addParam(qualifiedName);
-    pgm.addParam({ type: '10A', value: type });
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-
-          rtValue = {
-            'Object_authority_/_Data_authority': data[2]._,
-            Authorization_list_management: data[3]._,
-            Object_operational: data[4]._,
-            Object_management: data[5]._,
-            Object_existence: data[6]._,
-            Data_read: data[7]._,
-            Data_add: data[8]._,
-            Data_update: data[9]._,
-            Data_delete: data[10]._,
-            Authorization_list: data[11]._,
-            Authority_source: data[12]._,
-            Some_adopted_authority: data[13]._,
-            Adopted_object_authority: data[14]._,
-            Adopted_authorization_list_management: data[15]._,
-            Adopted_object_operational: data[16]._,
-            Adopted_object_management: data[17]._,
-            Adopted_object_existence: data[18]._,
-            Adopted_data_read: data[19]._,
-            Adopted_data_add: data[20]._,
-            Adopted_data_update: data[21]._,
-            Adopted_data_delete: data[22]._,
-            Adopted_data_execute: data[23]._,
-            Reserved: data[24]._,
-            Adopted_object_alter: data[25]._,
-            Adopted_object_reference: data[26]._,
-            // Reserved: data[27]._, // TODO: duplicate key
-            Data_execute: data[28]._,
-            // Reserved: data[29]._, // TODO: duplicate key
-            Object_alter: data[30]._,
-            Object_reference: data[31]._,
-            ASP_device_name_of_library: data[32]._,
-            ASP_device_name_of_object: data[33]._,
-            // Reserved: data[34]._, // TODO: duplicate key
-            Offset_to_group_information_table: data[35]._,
-            Number_of_group_table_entries_returned: data[36]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  retrCmdInfo(cmd, lib = '*LIBL', cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // Command name
-        { type: '10A', value: '' }, // Command library name
-        { type: '10A', value: '' }, // Command processing program or proxy target command
-        { type: '10A', value: '' }, // Command processing program's or proxy target command's library name
-        { type: '10A', value: '' }, // Source file name
-        { type: '10A', value: '' }, // Source file library name
-        { type: '10A', value: '' }, // Source file member name
-        { type: '10A', value: '' }, // Validity check program name
-        { type: '10A', value: '' }, // ]Validity check program library name
-        { type: '10A', value: '' }, // ]Mode information
-        { type: '15A', value: '' }, // ]Where allowed to run
-        { type: '1A', value: '' }, // Allow limited user
-        { type: '10i0', value: 0 }, // Maximum positional parameters
-        { type: '10A', value: '' }, // Prompt message file name
-        { type: '10A', value: '' }, // Prompt message file library name
-        { type: '10A', value: '' }, // Message file name
-        { type: '10A', value: '' }, // Message file library name
-        { type: '10A', value: '' }, // Help panel group name
-        { type: '10A', value: '' }, // Help panel group library name
-        { type: '10A', value: '' }, // Help identifier
-        { type: '10A', value: '' }, // Search index name
-        { type: '10A', value: '' }, // Search index library name
-        { type: '10A', value: '' }, // Current library
-        { type: '10A', value: '' }, // Product library
-        { type: '10A', value: '' }, // Prompt override program name
-        { type: '10A', value: '' }, // Prompt override program library name
-        { type: '6A', value: '' }, // Restricted to target release
-        { type: '50A', value: '' }, // Text description
-        { type: '2A', value: '' }, // Command processing program call state
-        { type: '2A', value: '' }, // Validity check program call state
-        { type: '2A', value: '' }, // Prompt override program call state
-        { type: '10i0', value: 0 }, // Offset to help bookshelf information
-        { type: '10i0', value: 0 }, // Length of help bookshelf information
-        { type: '10i0', value: 0 }, // Coded character set ID (CCSID)
-        { type: '1A', value: '' }, // Enabled for GUI indicator
-        { type: '1A', value: '' }, // Threadsafe indicator
-        { type: '1A', value: '' }, // Multithreaded job action
-        { type: '1A', value: '' }, // Proxy command indicator
-        { type: '1A', value: '' }, // Prompt message file text indicator
-        { type: '13A', value: '' }, // Reserved
-      ],
-    };
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: cmd },
-        { type: '10A', value: lib },
-      ],
-    };
-    const pgm = new ProgramCall('QCDRCMDI', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'CMDI0100' });
-    pgm.addParam(qualifiedName);
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-
-          rtValue = {
-            Command_name: data[2]._,
-            Command_library_name: data[3]._,
-            Command_processing_program_or_proxy_target_command: data[4]._,
-            "Command_processing_program's_or_proxy_target_command's_library_name": data[5]._,
-            Source_file_name: data[6]._,
-            Source_file_library_name: data[7]._,
-            Source_file_member_name: data[8]._,
-            Validity_check_program_name: data[9]._,
-            Validity_check_program_library_name: data[10]._,
-            Mode_information: data[11]._,
-            Where_allowed_to_run: data[12]._,
-            Allow_limited_user: data[13]._,
-            Maximum_positional_parameters: data[14]._,
-            Prompt_message_file_name: data[15]._,
-            Prompt_message_file_library_name: data[16]._,
-            Message_file_name: data[17]._,
-            Message_file_library_name: data[18]._,
-            Help_panel_group_name: data[19]._,
-            Help_panel_group_library_name: data[20]._,
-            Help_identifier: data[21]._,
-            Search_index_name: data[22]._,
-            Search_index_library_name: data[23]._,
-            Current_library: data[24]._,
-            Product_library: data[25]._,
-            Prompt_override_program_name: data[26]._,
-            Prompt_override_program_library_name: data[27]._,
-            Restricted_to_target_release: data[28]._,
-            Text_description: data[29]._,
-            Command_processing_program_call_state: data[30]._,
-            Validity_check_program_call_state: data[31]._,
-            Prompt_override_program_call_state: data[32]._,
-            Offset_to_help_bookshelf_information: data[33]._,
-            Length_of_help_bookshelf_information: data[34]._,
-            'Coded_character_set_ID_(CCSID)': data[35]._,
-            Enabled_for_GUI_indicator: data[36]._,
-            Threadsafe_indicator: data[37]._,
-            Multithreaded_job_action: data[38]._,
-            Proxy_command_indicator: data[39]._,
-            Prompt_message_file_text_indicator: data[40]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  retrPgmInfo(_pgm, lib = '*LIBL', cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // Program name
-        { type: '10A', value: '' }, // Program library name
-        { type: '10A', value: '' }, // Program owner
-        { type: '10A', value: '' }, // Program attribute
-        { type: '13A', value: '' }, // Creation date and time
-        { type: '10A', value: '' }, // Source file name
-        { type: '10A', value: '' }, // Source file library name
-        { type: '10A', value: '' }, // Source file member name
-        { type: '13A', value: '' }, // [10]Source file updated date and time
-        { type: '1A', value: '' }, // Observable information
-        { type: '1A', value: '' }, // User profile option
-        { type: '1A', value: '' }, // Use adopted authority
-        { type: '1A', value: '' }, // Log commands
-        { type: '1A', value: '' }, // Allow RTVCLSRC
-        { type: '1A', value: '' }, // Fix decimal data
-        { type: '50A', value: '' }, // Text description
-        { type: '1A', value: '' }, // Type of program
-        { type: '1A', value: '' }, // Teraspace storage-enabled program
-        { type: '58A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Minimum number of parameters
-        { type: '10i0', value: 0 }, // Maximum number of parameters
-        { type: '10i0', value: 0 }, // Program size
-        { type: '10i0', value: 0 }, // Associated space size
-        { type: '10i0', value: 0 }, // Static storage size
-        { type: '10i0', value: 0 }, // Automatic storage size
-        { type: '10i0', value: 0 }, // Number of MI instructions
-        { type: '10i0', value: 0 }, // Number of MI ODT entries
-        { type: '1A', value: '' }, // Program state
-        { type: '14A', value: '' }, // Compiler identification
-        { type: '6A', value: '' }, // Earliest release program can run
-        { type: '10A', value: '' }, // Sort sequence table name
-        { type: '10A', value: '' }, // Sort sequence table library name
-        { type: '10A', value: '' }, // Language identifier
-        { type: '1A', value: '' }, // Program domain
-        { type: '1A', value: '' }, // Conversion required
-        { type: '1A', value: '' }, // Conversion details
-        { type: '19A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // Optimization
-        { type: '1A', value: '' }, // Paging pool
-        { type: '1A', value: '' }, // Update program automatic storage area (PASA)
-        { type: '1A', value: '' }, // Clear program automatic storage area (PASA)
-        { type: '1A', value: '' }, // Paging amount
-        { type: '18A', value: '' }, // Reserved
-        { type: '10A', value: '' }, // Program entry procedure module
-        { type: '10A', value: '' }, // Program entry procedure module library
-        { type: '30A', value: '' }, // Activation group attribute
-        { type: '1A', value: '' }, // Observable information compressed
-        { type: '1A', value: '' }, // Run-time information compressed
-        { type: '6A', value: '' }, // Release program created on
-        { type: '1A', value: '' }, // Shared activation group
-        { type: '1A', value: '' }, // Allow update
-        { type: '10i0', value: 0 }, // Program CCSID
-        { type: '10i0', value: 0 }, // Number of modules
-        { type: '10i0', value: 0 }, // Number of service programs
-        { type: '10i0', value: 0 }, // Number of copyrights
-        { type: '10i0', value: 0 }, // Number of unresolved references
-        { type: '6A', value: '' }, // Release program created for
-        { type: '1A', value: '' }, // Allow static storage reinitialization
-        { type: '1A', value: '' }, // All creation data
-        { type: '1A', value: '' }, // Allow bound *SRVPGM library name update
-        { type: '10A', value: '' }, // Profiling data
-        { type: '1A', value: '' }, // Teraspace storage enabled modules
-        { type: '1A', value: '' }, // Storage model
-        { type: '10A', value: '' }, // Uses argument optimization (ARGOPT)
-        { type: '77A', value: '' }, // Reserved
-      ],
-    };
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: _pgm },
-        { type: '10A', value: lib },
-      ],
-    };
-    const pgm = new ProgramCall('QCLRPGMI', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'PGMI0100' });
-    pgm.addParam(qualifiedName);
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Program_name: data[2]._,
-            Program_library_name: data[3]._,
-            Program_owner: data[4]._,
-            Program_attribute: data[5]._,
-            Creation_date_and_time: data[6]._,
-            Source_file_name: data[7]._,
-            Source_file_library_name: data[8]._,
-            Source_file_member_name: data[9]._,
-            Source_file_updated_date_and_time: data[10]._,
-            Observable_information: data[11]._,
-            User_profile_option: data[12]._,
-            Use_adopted_authority: data[13]._,
-            Log_commands: data[14]._,
-            Allow_RTVCLSRC: data[15]._,
-            Fix_decimal_data: data[16]._,
-            Text_description: data[17]._,
-            Type_of_program: data[18]._,
-            'Teraspace_storage-enabled_program': data[19]._,
-            Reserved: data[20]._,
-            Minimum_number_of_parameters: data[21]._,
-            Maximum_number_of_parameters: data[22]._,
-            Program_size: data[23]._,
-            Associated_space_size: data[24]._,
-            Static_storage_size: data[25]._,
-            Automatic_storage_size: data[26]._,
-            Number_of_MI_instructions: data[27]._,
-            Number_of_MI_ODT_entries: data[28]._,
-            Program_state: data[29]._,
-            Compiler_identification: data[30]._,
-            Earliest_release_program_can_run: data[31]._,
-            Sort_sequence_table_name: data[32]._,
-            Sort_sequence_table_library_name: data[33]._,
-            Language_identifier: data[34]._,
-            Program_domain: data[35]._,
-            Conversion_required: data[36]._,
-            Conversion_details: data[37]._,
-            // Reserved: data[38]._, // TODO: duplicate
-            Optimization: data[39]._,
-            Paging_pool: data[40]._,
-            'Update_program_automatic_storage_area_(PASA)': data[41]._,
-            'Clear_program_automatic_storage_area_(PASA)': data[42]._,
-            Paging_amount: data[43]._,
-            // Reserved: data[38]._, // TODO: duplicate
-            Program_entry_procedure_module: data[45]._,
-            Program_entry_procedure_module_library: data[46]._,
-            Activation_group_attribute: data[47]._,
-            Observable_information_compressed: data[48]._,
-            'Run-time_information_compressed': data[49]._,
-            Release_program_created_on: data[50]._,
-            Shared_activation_group: data[51]._,
-            Allow_update: data[52]._,
-            Program_CCSID: data[53]._,
-            Number_of_modules: data[54]._,
-            Number_of_service_programs: data[55]._,
-            Number_of_copyrights: data[56]._,
-            Number_of_unresolved_references: data[57]._,
-            Release_program_created_for: data[58]._,
-            Allow_static_storage_reinitialization: data[59]._,
-            All_creation_data: data[60]._,
-            'Allow_bound_*SRVPGM_library_name_update': data[61]._,
-            Profiling_data: data[62]._,
-            Teraspace_storage_enabled_modules: data[63]._,
-            Storage_model: data[64]._,
-            'Uses_argument_optimization_(ARGOPT)': data[65]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  retrSrvPgmInfo(srvpgm, lib = '*LIB', cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // Service program name
-        { type: '10A', value: '' }, // Service program library name
-        { type: '10A', value: '' }, // Service program owner
-        { type: '10A', value: '' }, // Service program attribute
-        { type: '13A', value: '' }, // Creation date and time
-        { type: '10A', value: '' }, // Export source file name
-        { type: '10A', value: '' }, // Export source file library name
-        { type: '10A', value: '' }, // Export source file member name
-        { type: '30A', value: '' }, // ]Activation group attribute
-        { type: '16A', value: '' }, // ]Current export signature
-        { type: '1A', value: '' }, // User profile
-        { type: '1A', value: '' }, // Observable information compressed
-        { type: '1A', value: '' }, // Run-time information compressed
-        { type: '10i0', value: 0 }, // Service program CCSID
-        { type: '10i0', value: 0 }, // Number of modules
-        { type: '10i0', value: 0 }, // Number of service programs
-        { type: '10i0', value: 0 }, // Number of copyrights
-        { type: '50A', value: '' }, // Text description
-        { type: '1A', value: '' }, // Shared activation group
-        { type: '1A', value: '' }, // Allow update
-        { type: '10i0', value: 0 }, // [22]Number of unresolved references
-        { type: '1A', value: '' }, // Use adopted authority
-        { type: '1A', value: '' }, // Allow bound *SRVPGM library name update
-        { type: '10A', value: '' }, // Profiling data
-        { type: '1A', value: '' }, // Teraspace storage enabled modules
-        { type: '1A', value: '' }, // Storage model
-        { type: '10A', value: '' }, // Uses argument optimization (ARGOPT)
-        { type: '70A', value: '' }, // Reserved '00'X
-        { type: '1A', value: '' }, // Service program state
-        { type: '1A', value: '' }, // Service program domain
-        { type: '10i0', value: 0 }, // Associated space size
-        { type: '10i0', value: 0 }, // Static storage size
-        { type: '10i0', value: 0 }, // Service program size
-        { type: '6A', value: '' }, // Release service program created on
-        { type: '6A', value: '' }, // Earliest release service program can run
-        { type: '6A', value: '' }, // Release service program created for
-        { type: '1A', value: '' }, // Allow static storage reinitialization
-        { type: '1A', value: '' }, // Conversion required
-        { type: '1A', value: '' }, // All creation data
-        { type: '1A', value: '' }, // Conversion details
-        { type: '90A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // Paging pool
-        { type: '1A', value: '' }, // Paging amount
-      ],
-    };
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: srvpgm },
-        { type: '10A', value: lib },
-      ],
-    };
-    const pgm = new ProgramCall('QBNRSPGM', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'SPGI0100' });
-    pgm.addParam(qualifiedName);
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-
-          rtValue = {
-            Service_program_name: data[2]._,
-            Service_program_library_name: data[3]._,
-            Service_program_owner: data[4]._,
-            Service_program_attribute: data[5]._,
-            Creation_date_and_time: data[6]._,
-            Export_source_file_name: data[7]._,
-            Export_source_file_library_name: data[8]._,
-            Export_source_file_member_name: data[9]._,
-            Activation_group_attribute: data[10]._,
-            Current_export_signature: data[11]._,
-            User_profile: data[12]._,
-            Observable_information_compressed: data[13]._,
-            'Run-time_information_compressed': data[14]._,
-            Service_program_CCSID: data[15]._,
-            Number_of_modules: data[16]._,
-            Number_of_service_programs: data[17]._,
-            Number_of_copyrights: data[18]._,
-            Text_description: data[19]._,
-            Shared_activation_group: data[20]._,
-            Allow_update: data[21]._,
-            Number_of_unresolved_references: data[22]._,
-            Use_adopted_authority: data[23]._,
-            'Allow_bound_*SRVPGM_library_name_update': data[24]._,
-            Profiling_data: data[25]._,
-            Teraspace_storage_enabled_modules: data[26]._,
-            Storage_model: data[27]._,
-            'Uses_argument_optimization_(ARGOPT)': data[28]._,
-            "Reserved_'00'X": data[29]._,
-            Service_program_state: data[30]._,
-            Service_program_domain: data[31]._,
-            Associated_space_size: data[32]._,
-            Static_storage_size: data[33]._,
-            Service_program_size: data[34]._,
-            Release_service_program_created_on: data[35]._,
-            Earliest_release_service_program_can_run: data[36]._,
-            Release_service_program_created_for: data[37]._,
-            Allow_static_storage_reinitialization: data[38]._,
-            Conversion_required: data[39]._,
-            All_creation_data: data[40]._,
-            Conversion_details: data[41]._,
-            Reserved: data[42]._,
-            Paging_pool: data[43]._,
-            Paging_amount: data[44]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  retrUserInfo(user, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // User profile name
-        { type: '13A', value: '' }, // Previous sign-on date and time
-        { type: '1A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Sign-on attempts not valid
-        { type: '10A', value: '' }, // Status
-        { type: '8A', value: '' }, // Password change date
-        { type: '1A', value: '' }, // No password indicator
-        { type: '1A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Password expiration interval
-        { type: '8A', value: '' }, // Date password expires
-        { type: '10i0', value: 0 }, // Days until password expires
-        { type: '1A', value: '' }, // Set password to expire
-        { type: '10A', value: '' }, // Display sign-on information
-        { type: '1A', value: '' }, // Local password management
-        { type: '10A', value: '' }, // Block password change
-      ],
-    };
-    const pgm = new ProgramCall('QSYRUSRI', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'USRI0100' });
-    pgm.addParam({ type: '10A', value: user });
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            User_profile_name: data[2]._,
-            'Previous_sign-on_date_and_time': data[3]._,
-            // Reserved: data[4]._, // TODO: This is a duplicate
-            'Sign-on_attempts_not_valid': data[5]._,
-            Status: data[6]._,
-            Password_change_date: data[7]._,
-            No_password_indicator: data[8]._,
-            Reserved: data[9]._,
-            Password_expiration_interval: data[10]._,
-            Date_password_expires: data[11]._,
-            Days_until_password_expires: data[12]._,
-            Set_password_to_expire: data[13]._,
-            'Display_sign-on_information': data[14]._,
-            Local_password_management: data[15]._,
-            Block_password_change: data[16]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
 
-  retrUserAuthToObj(path, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10A', value: '' }, // Profile name
-        { type: '1A', value: '' }, // User or group indicator
-        { type: '10A', value: '' }, // Data authority
-        { type: '1A', value: '' }, // Authorization list management
-        { type: '1A', value: '' }, // Object management
-        { type: '1A', value: '' }, // Object existence
-        { type: '1A', value: '' }, // Object alter
-        { type: '1A', value: '' }, // Object reference
-        { type: '10A', value: '' }, // Reserved
-        { type: '1A', value: '' }, // bject operational
-        { type: '1A', value: '' }, // Data read
-        { type: '1A', value: '' }, // Data add
-        { type: '1A', value: '' }, // Data update
-        { type: '1A', value: '' }, // Data delete
-        { type: '1A', value: '' }, // Data execute
-        { type: '10A', value: '' }, // Reserved
-      ],
-    };
-    const feedBack = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec3',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned in the returned records feedback information
-        { type: '10i0', value: 0 }, // Bytes available in the returned records feedback information
-        { type: '10i0', value: 0 }, // Bytes returned in the receiver variable
-        { type: '10i0', value: 0 }, // Bytes available in the receiver variable
-        { type: '10i0', value: 0 }, // Number of authorized users
-        { type: '10i0', value: 0 }, // Entry length for each authorized user returned
-        { type: '10A', value: '' }, // Owner
-        { type: '10A', value: '' }, // Primary group
-        { type: '10A', value: '' }, // Authorization list
-        { type: '1A', value: '' }, // Sensitivity level
-      ],
-    };
 
-    const pgm = new ProgramCall('QSYRTVUA', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam(feedBack);
-    pgm.addParam({ type: '10i0', setlen: 'rec3', value: 0 });
-    pgm.addParam({ type: '8A', value: 'RTUA0100' });
-    pgm.addParam({ type: `${path.length}A`, value: path });
-    pgm.addParam({ type: '10i0', value: path.length });
-    pgm.addParam(this.errno);
-    this.conn.add(pgm.toXML());
 
-    let rtValue;
 
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Profile_name: data[0]._,
-            User_or_group_indicator: data[1]._,
-            Data_authority: data[2]._,
-            Authorization_list_management: data[3]._,
-            Object_management: data[4]._,
-            Object_existence: data[5]._,
-            Object_alter: data[6]._,
-            Object_reference: data[7]._,
-            Reserved: data[8]._,
-            Object_operational: data[9]._,
-            Data_read: data[10]._,
-            Data_add: data[11]._,
-            Data_update: data[12]._,
-            Data_delete: data[13]._,
-            Data_execute: data[14]._,
-          };
-        } else { rtValue = str; }
 
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
 
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
-  addToLibraryList(Lib, cb) {
-    const pgm = new ProgramCall('QLICHGLL', { lib: 'QSYS' });
-    pgm.addParam({ type: '11A', value: '*SAME' });
-    pgm.addParam({ type: '11A', value: '*SAME' });
-    pgm.addParam({ type: '11A', value: '*SAME' });
-    pgm.addParam({ type: '11A', value: Lib });
-    pgm.addParam({ type: '10i0', value: 1 });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = true;
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
   getPTFInfo(PTFID, cb) {
     const outBuf = {
